account_service/requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.19
redis[asyncio]==5.0.4
aio-pika==9.4.1
prometheus-client==0.20.0
pydantic-settings==2.2.1
